SwiftPay/requirements.txt
PyQt6>=6.4.0
mysql-connector-python>=8.0.0
bcrypt>=4.0.0
reportlab>=4.0.0
matplotlib>=3.7.0

